Add log to the example
In the single session mode of the dap, the IDE start the adapter as a child process and communicate with it using stdin and stdout. Since the dev tool is the parent process, it also capture stderr

As such, for debugging purpose, I added the machinery to log to a file.

It is taken strait from rust cookbook.