Add DAP event type
This commit add the type corresponding to the event define in the spec. Were the event reference a non event type, this commit leave the type as  a todo.

As this commit add the type, but does not add any code for using the type, I have added a #[allow(dead_code)] attribute.
It will be remove once I write the code to parse event.

Since I have doc base on the spec and the spec is license under the Creative Commons Attribution 3.0 United States License, I have added attribution.